Add files via upload
Hamlib is now working and I run JRX_TX from my terminal.  Talking to IC-7100 over its USB connection.  Have been making things modular and still want to put the Hamlib interface in a class "by itself".   Realize now the MacOs has not been tested for JRX because there are MacOS peculiarities that did not work with the code.  So patching things as we go and struggling with the lack of tools on the Mac such as usbmon (missing) and WireShark (does not work).  Will need a little app to sniff the communications so
we can debug.  NetBeans refuses to run the daemon rigcfgd.  Does not see the path or directory.  This means no debugger when we most need one.